floater

+++ jukebox/utils/io.py
@@ -49,7 +49,7 @@ def load_audio(file, sr, offset, duration, resample=True, approx=False, time_bas
             frame.pts = None
             frame = resampler.resample(frame)
         
-        frame = np.array(frame)
+        frame = np.array(frame).astype(float)
         print(frame)
         #frame = frame.to_ndarray(format='fltp') # Convert to floats and not int16
         read = frame.shape[-1]
